Adele-Backoffice/backend/requirements.txt
aiosmtpd==1.4.6
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
atpublic==6.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
fairclient==1.0.1
fastapi==0.115.12
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Mail==0.10.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
ipykernel==6.29.5
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
minio==7.2.15
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.19.1
PyJWT==2.4.0
pymongo==4.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pyzmq==26.4.0
rdflib==7.0.0
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
git+https://github.com/FracassandoCasualmente/SeMPyRO.git
six==1.17.0
sniffio==1.3.1
SPARQLWrapper==2.0.0
stack-data==0.6.3
starlette==0.46.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
Werkzeug==2.0.3
